+++ week3/haar_wavelet.m
@@ -9,5 +9,6 @@ function psi = haar_wavelet(x, j, k)
         psi = 2^(j / 2) * haar_wavelet(2^j * x - k);
         return;
     end
-    psi = (x >= 0 & x < 0.5) - (x >= 0.5 & x < 1);
+    % Conversion to double since we want numeric values, not boolean
+    psi = double((x >= 0 & x < 0.5)) - double((x >= 0.5 & x < 1));
 end
